Support for sonar selfhosted evidence
