+++ sonar-project.properties
@@ -1,9 +1,9 @@
 # Required metadata
-sonar.projectKey=test-evidence
+sonar.projectKey=evidence_test
 sonar.organization=my-evidence-test-org
-
+# sonar.host.url=
 # Project details
-sonar.projectName=test-evidence-project
+sonar.projectName=evidence_test
 sonar.projectVersion=1.0
 
 # Path to source directories
